Functionnal approch... by objects :'(
